+++ CHANGELOG.md
@@ -11,7 +11,7 @@
 - _Created_ `tooltip.lua` component.
 - Context-Menu entries now accept `Resolver Functions` that return dynamically generated values for context-menus during run-time.
 - _Created_ `ContextMenu:Add()` to handle table insertion properly. Can handle multiple entries at once.
-- _Added_ `TargetType` and `restrictTo` options.
+- _Added_ `TargetType` and `restrictUI` options.
 - _Created_ `Any::[Type]` activators for `Item` and `Characters`.
 
 ### CHANGED